fixing tests

+++ crates/anvil/tests/it/ipc/anvil_api.rs
@@ -280,7 +280,7 @@ async fn can_mine_manually() {
 #[tokio::test(flavor = "multi_thread")]
 #[serial]
 async fn test_set_next_timestamp() {
-    let (api, handle) = spawn(NodeConfig::test_ipc().with_fork_block_number(Some(0 as u64))).await;
+    let (api, handle) = spawn(NodeConfig::test_ipc().with_fork_block_number(Some(0u64))).await;
     let provider = handle.http_provider();
 
     let now = SystemTime::now().duration_since(SystemTime::UNIX_EPOCH).unwrap();
@@ -404,7 +404,7 @@ async fn test_timestamp_interval() {
 #[tokio::test(flavor = "multi_thread")]
 #[serial]
 async fn can_get_node_info() {
-    let (api, handle) = spawn(NodeConfig::test_ipc().with_fork_block_number(Some(0 as u64))).await;
+    let (api, handle) = spawn(NodeConfig::test_ipc().with_fork_block_number(Some(0u64))).await;
 
     let node_info = api.anvil_node_info().await.unwrap();
